requirements.txt
attrs==22.2.0
idna==3.4
iniconfig==2.0.0
mako==1.2.4
markdown==3.4.1
markupsafe==2.1.1
multidict==6.0.4
packaging==23.0
pdoc3==0.10.0
pluggy==1.0.0
pytest==7.2.0
pyyaml==6.0
six==1.16.0
vcrpy==4.2.1
wrapt==1.14.1
yarl==1.8.2
